refactor: Flatten module hierarchies in gallifrey, chronos, vortex, and shell
Follows the "Razor" philosophy to reduce unnecessary deep folder structures for small modules.

- Flattened `gallifrey/src/domain/` (`conversation.rs`, `knowledge.rs`, `state.rs`) into `gallifrey/src/domain.rs`.
- Flattened `gallifrey/src/query/mod.rs` into `gallifrey/src/query.rs`.
- Flattened `chronos/src/memory/mod.rs` into `chronos/src/memory.rs`.
- Flattened `vortex/src/model/` (`mod.rs`, `registry.rs`) into `vortex/src/model.rs`.
- Flattened `shell/src/experimental/` (`mod.rs`, `chronograph.rs`, `sonic.rs`) into `shell/src/experimental.rs`.
- Resolved associated `clippy` lints (`unwrap_used`, `panic`, `uninlined_format_args`, `doc_markdown`) in tests and implementations across the refactored crates.
