+++ chronos/src/experimental/doctor.rs
@@ -257,7 +257,12 @@ impl SystemDoctor {
 }
 
 #[cfg(test)]
-#[allow(clippy::unwrap_used)]
+#[allow(
+    clippy::unwrap_used,
+    clippy::unchecked_time_subtraction,
+    clippy::panic,
+    clippy::unused_async
+)]
 mod tests {
     use super::*;
     use std::collections::HashMap;
