+++ chronos/src/experimental/echoes.rs
@@ -88,6 +88,7 @@ impl EchoChamber {
 }
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used, clippy::panic, clippy::unused_async)]
 mod tests {
     use super::*;
     use tardis_common::id::{EntityId, ModelHandle, SessionId};
