+++ chronos/src/experimental/prophecy.rs
@@ -130,6 +130,7 @@ impl Prophet {
 }
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
     use serde_json::json;
