+++ chronos/src/experimental/psychic_paper.rs
@@ -215,6 +215,7 @@ impl PsychicPaper {
 }
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
     use serde_json::json;
