+++ gallifrey/src/experimental/entropy.rs
@@ -132,6 +132,7 @@ impl EntropyGauge {
 }
 
 #[cfg(test)]
+#[allow(clippy::unwrap_used)]
 mod tests {
     use super::*;
     use crate::Gallifrey;
