+++ gallifrey/src/stores/knowledge.rs
@@ -378,7 +378,7 @@ impl Default for KnowledgeStore {
 }
 
 #[cfg(test)]
-#[allow(clippy::unwrap_used)]
+#[allow(clippy::unwrap_used, clippy::panic)]
 mod tests {
     use super::*;
     use chrono::Utc;
