--- shell/src/experimental/mod.rs
@@ -1,3 +1,0 @@
-#[cfg(feature = "nova")]
-pub mod chronograph;
-pub mod sonic;
